fix: restructure pangochef stake sarstake swap state hooks
