+++ src/components/Pools/PangoChef/Stake/index.tsx
@@ -14,8 +14,8 @@ import { useApproveCallbackHook } from 'src/hooks/useApproveCallback';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
 import { usePairContract, usePangoChefContract } from 'src/hooks/useContract';
 import useTransactionDeadline from 'src/hooks/useTransactionDeadline';
-import { useHederaPangochefContractCreateCallback } from 'src/state/ppangoChef/hooks';
-import { usePangoChefStakeCallbackHook } from 'src/state/ppangoChef/multiChainsHooks';
+import { usePangoChefStakeCallbackHook } from 'src/state/ppangoChef/hooks';
+import { useHederaPangochefContractCreateCallback } from 'src/state/ppangoChef/hooks/hedera';
 import { PangoChefInfo } from 'src/state/ppangoChef/types';
 import { useDerivedStakeInfo, useGetPoolDollerWorth, useMinichefPendingRewards } from 'src/state/pstake/hooks';
 import { SpaceType } from 'src/state/pstake/types';
