+++ src/components/Pools/RemoveFarm/index.tsx
@@ -7,7 +7,7 @@ import { PNG } from 'src/constants/tokens';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
 import { useGetHederaTokenNotAssociated, useHederaTokenAssociated } from 'src/hooks/tokens/hedera';
-import { usePangoChefWithdrawCallbackHook } from 'src/state/ppangoChef/multiChainsHooks';
+import { usePangoChefWithdrawCallbackHook } from 'src/state/ppangoChef/hooks';
 import { useGetEarnedAmount, useGetRewardTokens, useMinichefPendingRewards } from 'src/state/pstake/hooks';
 import { StakingInfo } from 'src/state/pstake/types';
 import { useHederaPGLToken } from 'src/state/pwallet/hooks';
