+++ src/components/SarManageWidget/AddStake/index.tsx
@@ -15,8 +15,8 @@ import { useHederaTokenAssociated } from 'src/hooks/tokens/hedera';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
 import { useHederaSarNFTContract } from 'src/hooks/useContract';
 import { useWalletModalToggle } from 'src/state/papplication/hooks';
-import { useSarStakeInfo } from 'src/state/psarstake/hooks';
-import { useDerivativeSarStakeHook } from 'src/state/psarstake/multiChainsHooks';
+import { useDerivativeSarStakeHook } from 'src/state/psarstake/hooks';
+import { useSarStakeInfo } from 'src/state/psarstake/hooks/evm';
 import { Position } from 'src/state/psarstake/types';
 import { useTokenBalanceHook } from 'src/state/pwallet/multiChainsHooks';
 import { getBuyUrl } from 'src/utils';
