+++ src/components/SarManageWidget/Compound/index.tsx
@@ -11,8 +11,8 @@ import Tooltip from 'src/components/Tooltip';
 import { PNG } from 'src/constants/tokens';
 import { useChainId } from 'src/hooks';
 import { useUSDCPriceHook } from 'src/hooks/useUSDCPrice';
-import { useSarStakeInfo } from 'src/state/psarstake/hooks';
-import { useDerivativeSarCompoundHook } from 'src/state/psarstake/multiChainsHooks';
+import { useDerivativeSarCompoundHook } from 'src/state/psarstake/hooks';
+import { useSarStakeInfo } from 'src/state/psarstake/hooks/evm';
 import { Position } from 'src/state/psarstake/types';
 import ConfirmDrawer from '../ConfirmDrawer';
 import { Options } from '../types';
