+++ src/components/SwapWidget/SwapDetailInfo/index.tsx
@@ -4,9 +4,9 @@ import { useTranslation } from 'react-i18next';
 import { INITIAL_ALLOWED_SLIPPAGE, ONE_BIPS } from 'src/constants';
 import { BIPS_BASE } from 'src/constants/swap';
 import { Field } from 'src/state/pswap/actions';
+import { useDaasFeeInfo } from 'src/state/pswap/hooks/common';
 import { useUserSlippageTolerance } from 'src/state/puser/hooks';
 import { computeSlippageAdjustedAmounts, computeTradePriceBreakdown, warningSeverity } from 'src/utils/prices';
-import { useDaasFeeInfo } from '../../../state/pswap/hooks';
 import { Text } from '../../Text';
 import { ContentBox, DataBox, ValueText } from './styled';
