+++ src/components/SwapWidget/TWAP/TWAPPanel.tsx
@@ -6,8 +6,8 @@ import { SwapTypes } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { useAllTokens } from 'src/hooks/useAllTokens';
 import { useWalletModalToggle } from 'src/state/papplication/hooks';
-import { Field } from '../../../state/pswap/actions';
-import { useSwapActionHandlers } from '../../../state/pswap/hooks';
+import { Field } from 'src/state/pswap/actions';
+import { useSwapActionHandlers } from 'src/state/pswap/hooks/common';
 import SelectTokenDrawer from '../SelectTokenDrawer';
 import TradeOption from '../TradeOption';
 import { Root } from '../styled';
