+++ src/index.tsx
@@ -33,7 +33,7 @@ import ApplicationUpdater from 'src/state/papplication/updater';
 import { useCoinGeckoTokenData } from 'src/state/pcoingecko/hooks';
 import ListsUpdater from 'src/state/plists/updater';
 import MulticallUpdater from 'src/state/pmulticall/updater';
-import { usePangoChefInfosHook } from 'src/state/ppangoChef/multiChainsHooks';
+import { usePangoChefInfosHook } from 'src/state/ppangoChef/hooks';
 import {
   calculateTotalStakedAmountInAvax,
   calculateTotalStakedAmountInAvaxFromPng,
@@ -56,14 +56,13 @@ import {
   PoolType,
   StakingInfo,
 } from 'src/state/pstake/types';
+import { useGelatoLimitOrdersListHook } from 'src/state/pswap/hooks';
 import {
   LimitOrderInfo,
   useDerivedSwapInfo,
   useGelatoLimitOrderDetail,
   useSwapActionHandlers,
-} from 'src/state/pswap/hooks';
-
-import { useGelatoLimitOrdersListHook } from 'src/state/pswap/multiChainsHooks';
+} from 'src/state/pswap/hooks/common';
 
 import { useAllTransactions, useAllTransactionsClearer } from 'src/state/ptransactions/hooks';
 import TransactionUpdater from 'src/state/ptransactions/updater';
@@ -81,8 +80,8 @@ import { MixPanelEvents, MixPanelProvider, useMixpanel } from './hooks/mixpanel'
 import i18n, { availableLanguages } from './i18n';
 import store, { PANGOLIN_PERSISTED_KEYS, StoreContext, galetoStore, pangolinReducers } from './state';
 import { PangoChefInfo } from './state/ppangoChef/types';
-import { useSarStakeInfo } from './state/psarstake/hooks';
-import { useSarPositionsHook } from './state/psarstake/multiChainsHooks';
+import { useSarPositionsHook } from './state/psarstake/hooks';
+import { useSarStakeInfo } from './state/psarstake/hooks/evm';
 import { Position } from './state/psarstake/types';
 import SwapUpdater from './state/pswap/updater';
 import { default as ThemeProvider } from './theme';
