+++ src/state/pmint/hooks.ts
@@ -16,7 +16,7 @@ import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { AppState, useDispatch, useSelector } from 'src/state';
 import { PairState, usePair } from '../../data/Reserves';
 import { wrappedCurrency, wrappedCurrencyAmount } from '../../utils/wrappedCurrency';
-import { tryParseAmount } from '../pswap/hooks';
+import { tryParseAmount } from '../pswap/hooks/common';
 import { useCurrencyBalances } from '../pwallet/hooks';
 import { Field, typeInput } from './actions';
 import { initialKeyState } from './reducer';
