+++ src/state/ppangoChef/types.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from '@ethersproject/bignumber';
-import { Price } from '@pangolindex/sdk';
+import { CurrencyAmount, Price, TokenAmount } from '@pangolindex/sdk';
 import { MinichefStakingInfo } from '../pstake/types';
 
 export enum PoolType {
@@ -35,4 +35,17 @@ export interface PangoChefInfo extends MinichefStakingInfo {
   poolType: PoolType;
   poolRewardRate: BigNumber;
   lockCount: number | undefined;
+}
+
+export interface WithdrawData {
+  version?: number;
+  poolId: string | undefined;
+  stakedAmount: TokenAmount;
+  stakingRewardAddress?: string;
+}
+
+export interface PangoChefCompoundData {
+  poolId: string | undefined;
+  isPNGPool: boolean;
+  amountToAdd: CurrencyAmount | TokenAmount;
 }
