+++ src/state/psarstake/hooks/dummy.ts
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { PNG } from 'src/constants/tokens';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { ApprovalState } from 'src/hooks/useApproveCallback/constant';
-import { Position } from './types';
+import { Position } from '../types';
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 export function useDummySarPositions(): {
